+++ packages/lit-html/src/directive.ts
@@ -31,8 +31,10 @@ export type DirectiveParameters<C extends Directive> = Parameters<C['render']>;
  * returns a DirectiveResult object that captures the arguments.
  */
 export interface DirectiveResult<C extends DirectiveClass = DirectiveClass> {
-  /** @internal */
-  _$litDirective$: C;
+  /**
+   * This property needs to remain unminified.
+   * @internal */
+  ['_$litDirective$']: C;
   /** @internal */
   values: DirectiveParameters<InstanceType<C>>;
 }
@@ -82,7 +84,8 @@ export type PartInfo = ChildPartInfo | AttributePartInfo | ElementPartInfo;
 export const directive =
   <C extends DirectiveClass>(c: C) =>
   (...values: DirectiveParameters<InstanceType<C>>): DirectiveResult<C> => ({
-    _$litDirective$: c,
+    // This property needs to remain unminified.
+    ['_$litDirective$']: c,
     values,
   });
 
@@ -105,8 +108,9 @@ export abstract class Directive {
   // These will only exist on the AsyncDirective subclass
   //@internal
   _$disconnectableChildren?: Set<Disconnectable>;
+  // This property needs to remain unminified.
   //@internal
-  _$setDirectiveConnected?(isConnected: boolean): void;
+  ['_$setDirectiveConnected']?(isConnected: boolean): void;
 
   constructor(_partInfo: PartInfo) {}