+++ packages/lit-html/src/lit-html.ts
@@ -241,7 +241,8 @@ const COMMENT_PART = 7;
  * The return type of the template tag functions.
  */
 export type TemplateResult<T extends ResultType = ResultType> = {
-  _$litType$: T;
+  // This property needs to remain unminified.
+  ['_$litType$']: T;
   // TODO (justinfagnani): consider shorter names, like `s` and `v`. This is a
   // semi-public API though. We can't just let Terser rename them for us,
   // because we need TemplateResults to work between compatible versions of
@@ -257,7 +258,8 @@ export type SVGTemplateResult = TemplateResult<typeof SVG_RESULT>;
 export interface CompiledTemplateResult {
   // This is a factory in order to make template initialization lazy
   // and allow ShadyRenderOptions scope to be passed in.
-  _$litType$: CompiledTemplate;
+  // This property needs to remain unminified.
+  ['_$litType$']: CompiledTemplate;
   values: unknown[];
 }
 
@@ -274,9 +276,10 @@ export interface CompiledTemplate extends Omit<Template, 'el'> {
  * the given result type.
  */
 const tag =
-  <T extends ResultType>(_$litType$: T) =>
+  <T extends ResultType>(type: T) =>
   (strings: TemplateStringsArray, ...values: unknown[]): TemplateResult<T> => ({
-    _$litType$,
+    // This property needs to remain unminified.
+    ['_$litType$']: type,
     strings,
     values,
   });
@@ -355,8 +358,9 @@ export const render = (
   options?: RenderOptions
 ): ChildPart => {
   const partOwnerNode = options?.renderBefore ?? container;
+  // This property needs to remain unminified.
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let part: ChildPart = (partOwnerNode as any)._$litPart$;
+  let part: ChildPart = (partOwnerNode as any)['_$litPart$'];
   if (part === undefined) {
     // Internal modification: don't clear container to match lit-html 2.0
     if (
@@ -367,8 +371,9 @@ export const render = (
       container.childNodes.forEach((c) => c.remove());
     }
     const endNode = options?.renderBefore ?? null;
+    // This property needs to remain unminified.
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (partOwnerNode as any)._$litPart$ = part = new ChildPart(
+    (partOwnerNode as any)['_$litPart$'] = part = new ChildPart(
       container.insertBefore(createMarker(), endNode),
       endNode,
       undefined,
@@ -589,7 +594,8 @@ class Template {
   parts: Array<TemplatePart> = [];
 
   constructor(
-    {strings, _$litType$: type}: TemplateResult,
+    // This property needs to remain unminified.
+    {strings, ['_$litType$']: type}: TemplateResult,
     options?: RenderOptions
   ) {
     let node: Node | null;
@@ -748,9 +754,11 @@ function resolveDirective(
       : (parent as ChildPart | ElementPart | Directive).__directive;
   const nextDirectiveConstructor = isPrimitive(value)
     ? undefined
-    : (value as DirectiveResult)._$litDirective$;
+    : // This property needs to remain unminified.
+      (value as DirectiveResult)['_$litDirective$'];
   if (currentDirective?.constructor !== nextDirectiveConstructor) {
-    currentDirective?._$setDirectiveConnected?.(false);
+    // This property needs to remain unminified.
+    currentDirective?.['_$setDirectiveConnected']?.(false);
     if (nextDirectiveConstructor === undefined) {
       currentDirective = undefined;
     } else {
@@ -1010,7 +1018,8 @@ class ChildPart {
       } else if (value !== this._$committedValue && value !== noChange) {
         this._commitText(value);
       }
-    } else if ((value as TemplateResult)._$litType$ !== undefined) {
+      // This property needs to remain unminified.
+    } else if ((value as TemplateResult)['_$litType$'] !== undefined) {
       this._commitTemplateResult(value as TemplateResult);
     } else if ((value as Node).nodeType !== undefined) {
       this._commitNode(value as Node);
@@ -1090,20 +1099,18 @@ class ChildPart {
   private _commitTemplateResult(
     result: TemplateResult | CompiledTemplateResult
   ): void {
-    const {values, _$litType$} = result;
+    // This property needs to remain unminified.
+    const {values, ['_$litType$']: type} = result;
     // If $litType$ is a number, result is a plain TemplateResult and we get
     // the template from the template cache. If not, result is a
     // CompiledTemplateResult and _$litType$ is a CompiledTemplate and we need
     // to create the <template> element the first time we see it.
     const template: Template | CompiledTemplate =
-      typeof _$litType$ === 'number'
+      typeof type === 'number'
         ? this._$getTemplate(result as TemplateResult)
-        : (_$litType$.el === undefined &&
-            (_$litType$.el = Template.createElement(
-              _$litType$.h,
-              this.options
-            )),
-          _$litType$);
+        : (type.el === undefined &&
+            (type.el = Template.createElement(type.h, this.options)),
+          type);
 
     if ((this._$committedValue as TemplateInstance)?._$template === template) {
       (this._$committedValue as TemplateInstance)._update(values);
@@ -1232,12 +1239,6 @@ class AttributePart {
   _$disconnectableChildren?: Set<Disconnectable> = undefined;
 
   protected _sanitizer: ValueSanitizer | undefined;
-  /** @internal */
-  _setDirectiveConnected?: (
-    directive: Directive | undefined,
-    isConnected: boolean,
-    removeFromParent?: boolean
-  ) => void = undefined;
 
   get tagName() {
     return this.element.tagName;
@@ -1487,13 +1488,6 @@ class ElementPart {
 
   /** @internal */
   _$disconnectableChildren?: Set<Disconnectable> = undefined;
-
-  /** @internal */
-  _setDirectiveConnected?: (
-    directive: Directive | undefined,
-    isConnected: boolean,
-    removeFromParent?: boolean
-  ) => void = undefined;
 
   options: RenderOptions | undefined;